python_project/pyproject.toml
[project]
name = "bot-zap"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "elevenlabs (>=2.7.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "fastapi dev src/app.py"